Create README.md
we will import the required libraries. Next, we will load the housing data from the scikit-learn library and understand it. The prices of the house indicated by the variable MEDV is our target variable and the remaining are the feature variables based on which we will predict the value of a house.